Creating comment for posts moved to new component named "CreateCommentForm"
